// front/src/api/load_balancers/apply-clb/types.ts
export type NetworkAccountType = 'STANDARD' | 'LEGACY';
// ...
export type ExclusiveClusterType = 'TGW' | 'STGW';

export interface ExclusiveClusterItem {
  cloud_cluster_id: string;
  cluster_id: string;
  cluster_name: string;
  egress: string;
  isp: string;
  zone: string;
}

export interface ExclusiveClusterTag {
  cluster_tag: string;
  cluster_type: ExclusiveClusterType;
  clusters: ExclusiveClusterItem[];
}

// 独占集群标签查询支持的运营商，仅三网直连；BGP 等其它类型不支持独占集群
export type ExclusiveClusterIsp = 'CMCC' | 'CUCC' | 'CTCC';
// ...
export interface ApplyClbModel {
  // 业务ID
  bk_biz_id: number;
  // 账号ID
  account_id: string;
  // 地域
  region: string;
  // 网络类型: 公网 OPEN, 内网 INTERNAL
  load_balancer_type: 'OPEN' | 'INTERNAL';
  // 名称
  name: string;
  // 主可用区, 仅限公网型（自研云内网支持多可用区）
  zones: string | string[];
  // 备可用区, 目前仅广州、上海、南京、北京、中国香港、首尔地域的 IPv4 版本的 CLB 支持主备可用区。
  backup_zones?: string;
  // ip版本: IPV4, IPV6(ipv6 nat64), IPv6FullChain(ipv6)
  address_ip_version?: 'IPV4' | 'IPv6FullChain' | 'IPV6';
  // 安全组放通模式
  load_balancer_pass_to_target: boolean;
  // 云VpcID
  cloud_vpc_id: string;
  // 云子网ID, 内网型必填
  cloud_subnet_id?: string;
  // 绑定已有eip的ip地址, ipv6 nat64 不支持
  vip?: string;
  // 绑定eip id
  cloud_eip_id?: string;
  // 运营商类型(仅公网), 枚举值: CMCC, CUCC, CTCC, BGP。通过 TCloudDescribeResource 接口确定
  vip_isp?: string;
  // 网络计费模式(暂不支持包月)
  internet_charge_type?: 'TRAFFIC_POSTPAID_BY_HOUR' | 'BANDWIDTH_POSTPAID_BY_HOUR' | 'BANDWIDTH_PACKAGE';
  // 最大出带宽，单位Mbps
  internet_max_bandwidth_out?: number;
  // 带宽包id，计费模式为带宽包计费时必填
  bandwidth_package_id?: string;
  // 负载均衡规格类型: 性能容量型规格, 留空为共享型
  sla_type?: string;
  // 是否独占型：1是、0否
  exclusive?: 0 | 1;
  // 四层独占集群云上ID
  cloud_cluster_ids?: string[];
  // 七层独占集群标签
  cluster_tag?: string;
  // // 按月付费自动续费(暂不支持包月)
  // auto_renew?: boolean;
  // 购买数量
  require_count: number;
  // 备注
  memo?: string;
  // 可用区类型, 0: 单可用区 2: 主备可用区（仅前端使用）
  zoneType: '0' | '1';
  // 云厂商（仅前端使用）
  vendor: VendorEnum;
  // 用户网络类型
  account_type: NetworkAccountType;
  // 负载均衡规格类型, 0：共享型 1：性能容量型 2：独占型（仅前端使用）
  slaType: '0' | '1' | '2';
  // 四层/七层独占集群配置（仅前端使用）
  enable_l4?: boolean;
  enable_l7?: boolean;
  l4_cluster_tag?: string;
  l4_cluster_id?: string;
  l4_vip?: string;
  exclusive_cluster_tags?: ExclusiveClusterTag[];
  // 【自研云】是否开启直通特性
  zhi_tong?: boolean;
  // 【自研云】指定Tgw独占集群标签, 指定"ziyan_mianliu"时开启免流特性
  tgw_group_name?: string;
  [key: string]: any;
}